Factor out the code to cache hot java types.
